requirements/api.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6
openpyxl

# PDF + Text Parsing
PyPDF2>=3.0.1
pdfplumber>=0.10.3
regex>=2023.12.25

# ML Models
scikit-learn>=1.4.0
xgboost>=1.5.0

# NLP & Embeddings
sentence-transformers>=2.7.0
transformers>=4.36.0
torch>=2.0.0

# pour skills_extractor.py
spacy>=3.7.0

# Vector Search
faiss-cpu>=1.8.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# MLOps
mlflow>=2.8.0

# LLM
groq>=0.4.1

# Testing
pytest>=9.0.0
pytest-cov>=7.0.0